split test utils from adapters spec
